fix a typo
I'm not sure if fixing typos are considered an improvement, but this post is the first one in the search results when I type 'frontend github workflow' into Google search bar. So, I believe it must be perfect in every way... I frequently read your posts, thanks for the stuff you share, by the way.