Improve and Correct the English language
The & should not really be used as a shortcut for 'and' -- a lot of people do this, but it is ugly and not really correct. I almost never use the & sign, unless it is part of the code.

'via' is obviously a Latin word, and generally, good written English tries to avoid the use of Latin if it is not necessary. Some English language learners will not know what 'via' means.

I would probably normally write 'from the command line', but 'on the command line' is not wrong. However, I used 'on' because you use it later.

'kinda', as well as 'gonna', 'wanna' is not correct at all. This is how many people speak, but it should really be avoided in writing. 

I added "With the 'useCustom' hook" -- I think that is correct.

I did not replace 'caveat'. Again this is Latin. I would have to change the whole phrase, so I left it as it was.

People use 'pretty' instead of 'almost', 'really', etc. But it is generally wrong. 'pretty' means that something is nice or beautiful. I avoid using it in that way.

Why write four words when one will do? I changed 'Last but not least' to 'Finally', they really mean the same.