Added callback to the useEffect dependency array
Encountered a bug when I used the useOutsideClick hook. The state inside the callback doesn't update. Added the callback to the useEffect dependency array so the callback closure updates and you don't run a stale version of the function. 

Can test with this:

const [outsideClicks, setOutsideClicks] = useState(0)

const handleOutsideClick = () => {
    console.log('With/without callback in dependecy array', outsideClicks)
    setOutsideClicks(outsideClicks + 1)
}

const ref = useOutsideClick(handleOutsideClick);